fix bug with missing section title
Since the section title was put only for pages with a page_order of 0,
that meant the page title could be missed if the first image of a folder
was a portrait but the two following ones were landscapes.

This new implementation replaces the title placeholder once all pages
have been written, at which point we know the precise ordering.